Install npm in repo
