+++ package.json
@@ -1,0 +1,28 @@
+{
+  "name": "japt-2",
+  "version": "0.0.3",
+  "description": "The new and improved JavaScript golfing language",
+  "main": "index.js",
+  "directories": {
+    "doc": "docs",
+    "test": "test"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ETHproductions/japt-2.git"
+  },
+  "keywords": [
+    "japt",
+    "esolang",
+    "golfing-language"
+  ],
+  "author": "ETHproductions",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/ETHproductions/japt-2/issues"
+  },
+  "homepage": "https://github.com/ETHproductions/japt-2#readme"
+}
